+++ CHANGELOG.rst
@@ -11,6 +11,13 @@ TBR
   ``web_poet.HttpClient``.
 * Introduced ``web_poet.Meta`` to pass arbitrary information
   inside a Page Object.
+* removed support for Python 3.6
+* added support for Python 3.10
+* Backward Incompatible Change:
+
+    * ``ResponseData`` is now ``HttpResponse`` which has a new
+      specific attribute types like ``HttpResponseBody`` and
+      ``HttpResponseHeaders``.
 
 
 0.1.1 (2021-06-02)